refactor(ui): extract shared environment services helper
Move service extraction out of the environment page so monitoring and
other views can reuse the same service list mapping.
